feat(bonus) : minus free service

+++ srcs/requirements/wordpress/tools/entrypoint.sh
@@ -3,8 +3,12 @@
 CONFIG_PATH="/var/www/html/wordpress/wp-config.php"
 CONFIG_PATH_TMP="/var/www/html/wordpress/tmp_conf.php"
 
-cat $CONFIG_PATH_TMP | envsubst > $CONFIG_PATH
-rm -rf $CONFIG_PATH_TMP
 
-sleep 5
+if [ ! -f $CONFIG_PATH ]; then
+	cat $CONFIG_PATH_TMP | envsubst > $CONFIG_PATH
+	wp --allow-root core install  --url=$WORDPRESS_URL --title=$WORDPRESS_DB_NAME --admin_user=$WORDPRESS_DB_USER  --admin_password=$WORDPRESS_DB_PASSWORD  --admin_email=$WORDPRESS_EMAIL
+	wp --allow-root plugin install redis-cache --activate
+	wp --allow-root redis enable
+fi
+
 exec "$@"
